Fixed a problem on timespec was using 64 bit time and was causing wrong results on functions like clock_gettime. clock_settime now doesn't accept CLOCK_MONOTONIC since it isn't settable. Some small fixed for constructors change

+++ library/include/sys/types.h
@@ -76,8 +76,8 @@ typedef	quad_t *	qaddr_t;
 /* Iconv descriptor type */
 typedef void *_iconv_t;
 
-#define _CLOCK_T_ unsigned long long /* clock() */
-#define _TIME_T_ long long           /* time() */
+#define _CLOCK_T_ unsigned long     /* clock() */
+#define _TIME_T_ long               /* time() */
 #define _CLOCKID_T_ unsigned long
 #define _TIMER_T_ unsigned long
